refactor(http): Move graphql code out to file
